+++ Bangpa/AppDelegate.swift
@@ -91,7 +91,7 @@ extension AppDelegate: GIDSignInDelegate {
             }
             let name = user.profile.name
             let email = user.profile.email
-            let userInfo = ["name": name, "email": email]
+            let userInfo = ["name": name, "email": email, "provider": "google"]
             NotificationCenter.default.post(name: BangpaNotificationName.loginSuccess.value, object: nil, userInfo: userInfo)
         }
     }
